pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/vidore_benchmark"]

[project]
name = "vidore_benchmark"
dynamic = ["version"]
description = "Vision Document Retrieval (ViDoRe): Benchmark. Evaluation code for the ColPali paper."
authors = [
    { name = "Manuel Faysse", email = "manuel.faysse@illuin.tech" },
    { name = "Hugues Sibille", email = "hugues.sibille@illuin.tech" },
    { name = "Tony Wu", email = "tony.wu@illuin.tech" },
]
maintainers = [
    { name = "Manuel Faysse", email = "manuel.faysse@illuin.tech" },
    { name = "Tony Wu", email = "tony.wu@illuin.tech" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "accelerate>=0.34.0,<1.7.0",
    "bitsandbytes; sys_platform != 'darwin'",
    "configue>=5.0.0",
    "datasets>=2.19.1",
    "gcsfs>=2025.3.0",
    "GPUtil>=1.4.0,<2.0.0",
    "mteb>=1.16.3,<2",
    "numpy",
    "peft>=0.14.0,<0.16.0",
    "pillow>=10.0.0,<11.3.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "requests",
    "scipy",
    "sentence-transformers>=3.0.1,<4.0.0",
    "sentencepiece>=0.2.0,<1.0.0",
    "skypilot[gcp]>=0.9.3",
    "tenacity>=8.0.0",
    "timm>=1.0.19",
    "torch>=2.6.0,<2.7.0",
    "transformers>=4.51.1,<4.52.0",
    "typer>=0.15.1",
    "wandb>=0.21.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.1", "ruff>=0.4.5"]

bge-m3 = ["FlagEmbedding>=1.2.10,<2.0.0"]
bm25 = ["nltk>=3.8.1,<4.0.0", "rank-bm25>=0.2.2,<1.0.0"]
cohere = ["cohere>=5.11.2,<6.0.0"]
colpali-engine = ["colpali-engine>=0.3.8,<0.4.0"]
dse = ["qwen-vl-utils==0.0.8"]
jina-clip = ["timm>=1.0.7,<2.0.0"]
siglip = ["protobuf>=4.25.3"]
all-retrievers = [
    "vidore-benchmark[bge-m3]",
    "vidore-benchmark[bm25]",
    "vidore-benchmark[cohere]",
    "vidore-benchmark[colpali-engine]",
    "vidore-benchmark[dse]",
    "vidore-benchmark[jina-clip]",
    "vidore-benchmark[siglip]",
]

all = ["vidore-benchmark[all-retrievers]", "vidore-benchmark[dev]"]

skypilot = ["skypilot>=0.9.3,<1.0.0"]

[project.urls]
homepage = "https://github.com/illuin-tech/vidore-benchmark"

[project.scripts]
vidore-benchmark = "vidore_benchmark.cli.main:app"

[tool.pytest.ini_options]
filterwarnings = ["ignore::Warning"]
markers = ["slow: marks test as slow"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
